libkbfs: obfuscate file and dir names in logs
This commit changes the interface for `dirData` and `KBFSOps` to only
allow for `data.PathPartString` instances instead of raw strings for
file and directory names.  Doing so ensures, at a compiler level, that
if log obfuscation is turned on, then those names will be logged as
obfuscated instead of in plaintext.

This requires a lot of changes:
* The `NodeCache` creates a new obfuscator for every new `Node`.
* Each `data.Path` needs to have a `ChildObfuscator`, so that
  `dirData` can use it to obfuscate all the child names of the
  directory.
* Every `op` needs to have a final path set, so the path's obfuscator
  can be used on the plaintext filenames stored in the op.  Note that
  we can't just use `data.PathPartString`s in the op's fields
  directly, because `op` is a data structure that is marshaled
  directly into bytes that are stored on the server.
* We no longer log file paths when sending notifications up to the
  service; these are just plaintext strings so they'd be hard to
  obfuscate, and it doesn't seem worthwhile.

This commit also obfuscates filenames in `libfs.FS`.

Note that in some places, especially during conflict resolution, it is
very hard making sure that a consistent obfuscator is used for each
directory, and sometimes it just wasn't possible and we make a new
obfuscator for a directory that might have another obfuscator floating
around somewhere.  In practice this means that obfuscation collisions
in the directory could use different suffixes ("-2") for different
files.  This seems like it should be so rare that it won't matter, but
it's something to be aware of.

This commit does NOT do the following things, which will be handled in
subsequent work:

* Obfuscate filenames logged directly by libfuse/libdokan/simplefs.
* Obfuscate symlink targets.
* Obfuscating any filenames logged by the service or GUI.

Issue: HOTPOT-60